Change parameter type of ValueStream.skip from int to long
More than 2G positions can legitimately be skipped at once

+++ presto-orc/src/main/java/com/facebook/presto/orc/stream/BooleanStream.java
@@ -77,7 +77,7 @@ public class BooleanStream
     }
 
     @Override
-    public void skip(int items)
+    public void skip(long items)
             throws IOException
     {
         if (bitsInData >= items) {
